+++ src/main/java/cn/edith/demo/community/mapper/QuestionMapper.java
@@ -24,13 +24,13 @@ public interface QuestionMapper {
     Integer count();
 
     @Select("select * from question  where creator = #{id}  limit #{offset}, #{size}")
-    List<Question> listByuserId(@Param(value = "id") Integer id, @Param(value = "offset") Integer offset, @Param(value = "size") Integer size);
+    List<Question> listByuserId(@Param(value = "id") Long id, @Param(value = "offset") Integer offset, @Param(value = "size") Integer size);
 
     @Select("select count(1) from question where creator = #{id}")
-    Integer countByUserId(@Param(value = "id") Integer id);
+    Integer countByUserId(@Param(value = "id") Long id);
 
     @Select("select * from question  where id = #{id} " )
-    Question getById(@Param(value = "id") Integer id);
+    Question getById(@Param(value = "id") Long id);
 
 
     @Update("update question set title = #{title},description = #{description} , gmt_modified = #{gmtModified},tag = #{tag} where id = #{id}" )
